% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout.r
\name{F.r.norm}
\alias{F.r.norm}
\title{Double truncated normal CDF}
\usage{
F.r.norm(x, sigma, xtrunc, mean = 0)
}
\arguments{
\item{x}{value of normal random variable at which to evaluate the CDF}

\item{sigma}{standard deviation of distribution}

\item{xtrunc}{x value at which to left- and right-trunctate}

\item{mean}{mean of distribution (default=0)}
}
\description{
Calculates CDF of left- and right-truncated normal distribution
}
\examples{
sigma = 1
x = seq(-6,6,length=200)
Fx.r = F.r.norm(x,sigma,xtrunc=2*sigma,mean=0)
plot(x,Fx.r,type="l")
lines(x,pnorm(x,mean=0,sd=sigma),lty=2,col="red")
Fx.r = F.r.norm(x,sigma,xtrunc=3*sigma,mean=0)
plot(x,Fx.r,type="l")
lines(x,pnorm(x,mean=0,sd=sigma),lty=2,col="red")

}
