% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout.r
\name{F.r.unif}
\alias{F.r.unif}
\title{Double truncated uniform CDF}
\usage{
F.r.unif(x, sigma, xtrunc, mean = 0)
}
\arguments{
\item{x}{value of normal random variable at which to evaluate the CDF}

\item{sigma}{half-width of uniform distribution}

\item{xtrunc}{x value at which to left- and right-trunctate}

\item{mean}{mean of distribution (default=0)}
}
\description{
Calculates CDF of left- and right-truncated uniform distribution
}
\examples{
sigma = 1
x = seq(-6,6,length=200)
Fx.r = F.r.unif(x,sigma,xtrunc=2*sigma,mean=0)
plot(x,Fx.r,type="l")
lines(x,punif(x,min=-sigma,max=sigma),lty=2,col="red")
Fx.r = F.r.unif(x,sigma,xtrunc=3*sigma,mean=0)
plot(x,Fx.r,type="l")
lines(x,pnorm(x,min=-sigma,max=sigma),lty=2,col="red")
}
