% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomega.r
\name{make.inout.tpm}
\alias{make.inout.tpm}
\title{Make in-out transition probability matrix}
\usage{
make.inout.tpm(sigma, dmax, w, nx = 500)
}
\arguments{
\item{sigma}{standard deviation of nomal location distribution}

\item{dmax}{maximum distance can move}

\item{w}{half-width of strip}

\item{nx}{number of intervals to use for approximate integral.}
}
\description{
Makes the transition probability matrix for in-out movement from strip. Position 
 after movement is assumed to be normal with mean equal to #'  starting 
 coordinate and standared deviation \code{sigma}.
}
\examples{
w = 1
sigma = 0.1
dmax = 5*sigma
tpm = make.inout.tpm(sigma,dmax,w)
apply(tpm,1,sum) # check that rows sum to 1

}
