% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsum.r
\name{nch}
\alias{nch}
\title{Calculates number of possible capture histories}
\usage{
nch(n1, n2 = n1)
}
\arguments{
\item{n1}{The number of observer 1 detections.}

\item{n2}{The number of observer 2 detections.}
}
\description{
Calculates number of possible capture histories, given the number of 
 detections by each observer.
}
