% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomega.r
\name{p.i2o}
\alias{p.i2o}
\title{Prob goes from inside strip to outside strip}
\usage{
p.i2o(sigma, dmax, w, nx = 500)
}
\arguments{
\item{sigma}{standard deviation of nomal location distribution}

\item{dmax}{maximum distance can move}

\item{w}{half-width of strip}

\item{nx}{number of intervals to use for approximate integral.}
}
\description{
Returns the transition probability from inside strip to outside strip. 
 Starting coordinate outside strip is assumed to be a uniform random variable 
 on \code{(-w,w)}, where \code{w} is half-width of strip and cooridnate zero 
 is centre of strip. Position after movement is assumed to be normal with 
 mean equal to starting coordinate and standared deviation \code{sigma}. 
 
 Deals only with positive initial coordinate as problem is symmetric about zero. 
 Integrates CDF of normal distribution only inside strip, from initial coordinate 
 \code{0} to \code{w}. Integration is done by trapezoid rule.
}
\examples{
w = 1
sigma = 0.1
dmax = 5*sigma
p.i2o(sigma,dmax=dmax,w)
}
