% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomega.r
\name{p.o2o}
\alias{p.o2o}
\title{Prob goes from outside strip to outside strip}
\usage{
p.o2o(sigma, dmax, w, nx = 500)
}
\arguments{
\item{sigma}{standard deviation of nomal location distribution}

\item{dmax}{maximum distance can move}

\item{w}{half-width of strip}

\item{nx}{number of intervals to use for approximate integral.}
}
\description{
Returns the transition probability from outside strip to outside strip. 
 Starting coordinate outside strip is assumed to be a uniform random variable 
 on \code{(w,w+dmax)}, where \code{w} is half-width of strip, \code{dmax} is maximum 
 distance from strip to consider, and cooridnate zero is centre of strip. Position 
 after movement is assumed to be normal with mean equal to starting coordinate 
 and standared deviation \code{sigma}. 
 
 Deals only with positive initial coordinate as problem is symmetric about zero.
 Integrates CDF of normal distribution only outside strip, from initial coordinate 
 \code{w} to \code{w+dmax}. Integration is done by trapezoid rule.
}
\examples{
w = 1
sigma = 0.1
dmax = 5*sigma
p.o2o(sigma,dmax=dmax,w)
}
