% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pomega.r
\name{p.t}
\alias{p.t}
\title{Calculates the probabilitis of the observable capture histories}
\usage{
p.t(E1, Ec, p, sigmarate, k, dmax.t, planespd, halfw.dist = NULL,
  adj.mvt = FALSE, io = TRUE, idbn = NULL, nts = 200,
  ft.norm = TRUE)
}
\arguments{
\item{E1}{expected time available per dive cycle}

\item{Ec}{expected dive cycle length}

\item{p}{vector of probs that each obs detects available animal in searched strip}

\item{sigmarate}{Brownian movement rate parameter}

\item{k}{time lag between observers}

\item{dmax.t}{maximum distance can move (in plane-seconds)}

\item{planespd}{speed observers move along transect}

\item{halfw.dist}{half-width of searched strip}

\item{adj.mvt}{if TRUE, uses Brownian hitting time PDF to average over in-out TPMs.}

\item{nts}{number of integration points for averaging over in-out TPMs.}

\item{ft.norm}{if TRUE uses normal to approximate Brownian hitting times, else 
uses exact expression for Brownian hitting times.}
}
\description{
Calculates the probabilitis of the three observable capture histories for animals available 
 according to Markov process and allowed to move in and out of searched
 strip according to Brownian motion.
}
