% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsum.r
\name{sigmapalm2sigmarate}
\alias{sigmapalm2sigmarate}
\title{Converts between Palm sigma and LCE sigma}
\usage{
sigmapalm2sigmarate(sigma, lag)
}
\arguments{
\item{lag}{The time lag.}

\item{sigmarate}{The Palm distibution sigma}
}
\description{
Converts between LCE sigma and Palm sigma.
}
