% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsum.r
\name{sigmarate2sigmapalm}
\alias{sigmarate2sigmapalm}
\title{Converts between Palm sigma and LCE sigma}
\usage{
sigmarate2sigmapalm(sigmarate, lag)
}
\arguments{
\item{sigmarate}{The Brownian motion with movement rate paramter}

\item{lag}{The time lag.}
}
\description{
Converts between LCE sigma and Palm sigma.
}
