% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsum.r
\name{speed2sigmarate}
\alias{speed2sigmarate}
\title{Brownian motion sigma from average speed and time}
\usage{
speed2sigmarate(speed, lag)
}
\arguments{
\item{lag}{The time over which the average speed is required.}

\item{sigmarate}{The Brownian motion with movement rate paramter}
}
\description{
Calculates the Brownian motion with movement rate paramter from the average 
 speed E(displacement/time) of a particle doing Brownian motion over a time
 period \code{lag}. 
 Uses the fact that the expected value of a Chi random variable is sqrt(2)/gamma(0.5).
 NB: Assumes that speed is in km/sec and returns Brownian motion rate paramter in m/sec.
}
