\name{twoplane-package}
\alias{twoplane-package}
\alias{twoplane}
\docType{package}
\title{
Mark-recapture distance sampling with unknown recaptures
}
\description{
Estimates density from mark-recapture line transect survey without any recapture (duplicate) identification. Does this on the basis of the locations of detections by each observer, an animal movement model and a Markov model for the animal availability process.
}
\details{
\tabular{ll}{
Package: \tab twoplane\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-02\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
David Borchers and Peter Nightingale

Maintainer: David Borchers <dlb@st-andrews.ac.uk>
}
\references{
Hiby, L. and Lovell, P. (1998) Using Aircraft in Tandem Formation to Estimate Abundance of Harbour Porpoise. Biometrics 54: 1280-1289.
}
\keyword{ package 
}
\seealso{
~~ Optional links to other man pages, e.g. \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
